refactor: useShowArrow 훅 생성
